fix(m1): CI cargo check E0425/E0277/E0597 + unused_mut
- pty/mod.rs: transcript_path iç blok dışına taşındı (E0425 — start_output_pump'a aktarım)
- db.rs finalize_task: u64 → i64 cast (rusqlite ToSql u64 desteklemez — E0277)
- runtime/mod.rs: state+guard aynı iç scope'ta (E0597 — monitor + pump)
- pty/mod.rs transcript_append_session_buffer: gereksiz mut kaldırıldı (unused_mut)

+++ src-tauri/src/db.rs
@@ -537,7 +537,8 @@ impl AppDb {
                 spent_tokens_output = spent_tokens_output + ?5,
                 completed_at = COALESCE(completed_at, datetime('now'))
          WHERE id = ?1",
-        params![task_id, column, cost, tokens_in, tokens_out],
+        // rusqlite ToSql: u64 desteklenmez → i64'e cast (E0277).
+        params![task_id, column, cost, tokens_in as i64, tokens_out as i64],
       )
       .map_err(|e| e.to_string())?;
     if updated == 0 {
